requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
dropbox==12.0.2
google-api-core==2.19.1
    # via google-api-python-client
google-api-python-client==2.137.0
google-auth==2.32.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.7
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
ply==3.11
    # via stone
proto-plus==1.24.0
    # via google-api-core
protobuf==5.27.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pyparsing==3.1.2
    # via httplib2
requests==2.32.3
    # via
    #   dropbox
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   dropbox
    #   stone
stone==3.3.1
    # via dropbox
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.2
    # via requests
